First pass at a hyper-schema test format.
This attempts to create a test suite format for hyper-schema,
based on the following high-level requirements:

* Find all applicable links, and recongnize if they are usable
* Correctly identify link context and target(s)
* Recognize link relation types that call for special handling
* Recognize what kinds of input are possible (protocol-independent)

There are other requirements that may get added as they are
clarified in the specification.  In particular, there are probably
optional HTTP-specific features involving "targetHints" and/or
"requestSchema" that should be tested as well.  But possibly
separately.